Bridge C++ listbox system messages

+++ GeneralsRust/Code/GameEngine/GameClient/src/gui/callbacks/download_menu.rs
@@ -19,8 +19,8 @@ use game_network::download_manager::{
 use gamelogic::helpers::TheGameLogic;
 use gamelogic::helpers::TheGameText;
 
-const KEY_ESC: u32 = 0x1B;
-const KEY_STATE_UP: u32 = 0x0001;
+const KEY_ESC: usize = 0x1B;
+const KEY_STATE_UP: usize = 0x0001;
 
 #[derive(Default)]
 struct DownloadMenuState {
