+++ GeneralsRust/Code/GameEngine/GameClient/src/gui/callbacks/popup_join_game.rs
@@ -12,8 +12,8 @@ use std::cell::RefCell;
 use std::rc::Rc;
 use std::sync::{Mutex, OnceLock};
 
-const KEY_ESC: u32 = 0x1B;
-const KEY_STATE_UP: u32 = 0x0001;
+const KEY_ESC: usize = 0x1B;
+const KEY_STATE_UP: usize = 0x0001;
 
 #[derive(Default)]
 struct PopupJoinState {
